Fixed audio settings/saving
I switched to using a singleton for all canvases grouped under one parent. This makes it so all the interlocking parts controlling audio/settings/saving/opening and closing the settings canvas are all persisted across all scenes. I still need to make a system that tells the canvas group which canvases to have active/unactive depending on which scene is being loaded